+++ packages/orq-rc/RELEASES.md
@@ -6079,3 +6079,13 @@ Based on:
 - [python v4.5.0-rc.2] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.2 - packages/orq-rc
+
+## 2026-02-24 18:44:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.728.0 (2.839.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.3 - packages/orq-rc